Corrected bugs in UIMA indexing. Final index shows slight performance improvements.
I re-created the CT index with the "ct-indexer" pipeline.
The resulting index is approx. 400MB larger than the original
clinicaltrials_v3 index, the reason is unknown. I did not research
this for know.
The results are a bit better on the 2018 data: hpictall now has infNDCG 0.5563,
hpictphrase has Rprec 0.4098 and P@10 0.558 (as apposed to 0.5545,
0.4089 and 0.542).