Removing the dash deletion from the CT XML handler.
I basically just want to check if that has an impact. I could imagine
so because gene names often contain dashes. However, the genes in the
topics are all written without a dash and without a whitespace. Thus,
the dash removal treatment might even help with gene matches.
Checking now.