Began to refactor the retrieval out of the ExperimentsBuilder.
For the LtR experiments, we need a more fine-grained retrieval
where we can redefine the topics for which result should be retrieved
in a flexible way. An Experiment reads all topics internally
and iterates over all of them and directly evaluates them.
I want to separate the experiment from the retrieval so that
the retrieval can be used alone.