Integrated sample qrels into the AggregatedGoldStandard.
Also moved originally Trec-specific code into the
AggregatedGoldStandard super class and parametrized it with
functions to get the required information from the qrel files.

Added quite some direct setting possibilities to the experimenter
to make it more flexible for the cross evaluation. Not sure if
the class should stay this way in the long run, for now it's working fine.

But there are now a number of new switches that can get confusing.
Perhaps the best example is the need for cross-corpus topic IDs.
Those need to be the same in the qrels and the results file. The qrels
file is written by the AggregatedGoldStandard, the results file by the
Retrieval. Both need to have set the same topic ID generator.

Also, while we had cross-dataset topic IDs  for quite some commits now,
the `sample_eval`
program complained about non-numeric topic identifiers. It still
printed out performance values so I don't know if this was an issue at
all. But to be on the safe side, I changed the topic identifiers
to numeric by using the challenger year as a specifier.