Adding the UIMA components to create indices with different similarities.
To create different indices, the `trec2018-pubmed-indexer` and the
`ct-indexer` are the components to run. The new class to create
all the different indices has to be run first, to create the
correct mappings with the different similarities.
What I did then is to create one CoStoSys subset table for each
similarity for both datasets, PubMed and CT, resulting in 14
subsets and 14 indices.
I then ran the pipeline, PubMed, for example, for the BM25 subset table
indexing into the BM25 index. I waited until the pipeline was completely
initialized, so that the reader and consumer descriptors (which hold
the information from where to read and where to index), changed both
information to the next similarity, run, waited for the initialization
and so on. One could also duplicate the pipelines with the different
settings and run them all at once, if desired.
When doing it my way it is important to edit the descriptors in the
`desc/` directory of the respective pipeline because this is the
directory from where the descriptors are actually read when running
the pipeline. The `descAll` directory is only used by the
pipeline builder so that also inactive components are seen.