Refactored all caches to the CacheAccess abstraction.
The idea is to have a local cache - as it was before - but also
a CacheServer which can handle multiple cache users across the network.
This should help to get more cache hits.
The application is supposed to use one cache type - local or remote -
at one time, not both at the same time. They are alternatives to
each other.