Make gold standards encapsulate a single, in-memory, data model
Sample qrels are a superset of qrels, so there's no need to keep both around. Moreover, the `DocumentList` in a GS can be changed in memory, at which point it turns out-of-sync with the underlying file.

Change constructors so that they receive a single file and load the data from it once. When needed, a file can be created using `writeSampleQrelFile`, when data will be synced.

When a traditional qrel file is needed (e.g. statistics or 2017 CT), use `writeQrelFile` instead. Add a helper method `isSampleGoldStandard` to aid on that.