Make `Retrieval` year-independent
`Retrieval` is reusable, so the year depends on the topics being queried. However, the topics can span multiple years in LTR, so there's not even a single year.

Moreover, year is used only for the experimentID (and thus for filename), so it should not have an impact unless two identical experiments with different years are run together. In that case, it's just better to name the experiment.