Resolves #41.
`Result` instances now have the map of returned `_source` fields and
specific methods to return the treatment field contents. This is, of
course, TREC-PM specific and probably be refactored by creating a
`Result` subclass disclosing those methods in the future.
Important is that ElasticSearch must be told which fields to extract
from the source at all.
This is done with `Retrieval#withStoredFields()`. The new
`StoredFieldsRegistry` class offers a method to get the required
fields based on `Challenge`, `Task` and `year`.
For the TREC-PM BA 2019 track, the appropriate call has been added to
`LiteratureArticlesRetrievalRegistry`.