Reproduced best performance with a new index.
The last commit had been tested on the old ´sigir19_v1.0` index. This
index does not have the keywords field. The index `trecpm18_pm` on
the emergency ES5.4 instance has the field. But this drops performance
to infndcg ~0.554 or something. Thus, the field is removed in this commit.
This leads to the exact same performance of infndcg of 0.5648 as the
last commit. The important thing is to add the genes.