Replacing most `File` references in code - to templates or resource files - with simple strings.
The strings are just the exact same paths to the template or resource files as before.
However, they are no longer resolved to files after a classpath lookup. Instead, the code
places where the templates or resources are actually read, now use a mechanism to look
the file up as classpath resource or as an actual file.
The main reason for this is to allow the framework to be packaged in a large JAR and still
find the resources. With the old file-resolution approach, this wouldn't work because the
JAR-internal files would be found but the resulting file paths (e.g. file://Users/faessler/trec-pm.jar!/config.xml, note the '!') were not usable in `FileInputStreams` or similar mechanisms.
This is all working now.