Checked TRECPM2019 BA results on clean, new-built index. Slight differences.
The BA index that we used had 42,603 documents that did not belong into this index but to the 2017/18 data due to a fluke in the indexing process (yes, I mean I made a mistake and had the wrong database connection configured and noted only after the processing had already begun). This was known to Ellen who said that invalid documents would just be ignored.
The performance on the new index is:
[TrecMetricsCreator  ] Got NDCG = 0.6828, infNDCG = 0.5774, P@5 = 0.685, P@10 = 0.655, P@15 = 0.635, R-Prec = 0.3572, set_recall = 0.867 for collection jlpmcommon2 with decorators ConditionalCancer_FileBased_DiseaseSynonym_DiseasePreferredTerm_GeneSynonym_WordRemoval_SubTemplate(/templates/biomedical_articles/jlpmcommon2.json)_ElasticSearchQuery

The performance on the original index was:
[TrecMetricsCreator  ] Got NDCG = 0.6827, infNDCG = 0.5783, P@5 = 0.685, P@10 = 0.6525, P@15 = 0.6333, R-Prec = 0.3572, set_recall = 0.867 for collection jlpmcommon2 with decorators ConditionalCancer_FileBased_DiseaseSynonym_DiseasePreferredTerm_GeneSynonym_WordRemoval_SubTemplate(/templates/biomedical_articles/jlpmcommon2.json)_ElasticSearchQuery

I will now work on the new index because it has the correct document set.