Finished the generic optimization templates, the SMAC configuration and the parameter reading.
Tested on the TRECPM BA 2019 task, the performance of the default SMAC configuration is
19:46:45  INFO [TrecMetricsCreator  ] Got NDCG = 0.6647, infNDCG = 0.5523, P@5 = 0.675, P@10 = 0.6425, P@15 = 0.5967, R-Prec = 0.3641, set_recall = 0.8293 for collection jlpmcommon2gen with decorators FeatureControlledConditionalCancer_DiseaseUmlsSynonym_DiseaseUmlsPreferredTerm_FileBased_GeneSynonym_WordRemoval_SubTemplate(/templates/biomedical_articles_generic/jlpmcommon2generic.json)_StaticMap([1, 0.1, 1.0, 1.0, phrase, 1.0, 1.0, 2.0, 1.0, phrase, 0.6, 1.0, 0.7, best_fields, 0.1, 1, gefitinib prevent prophylaxis targets efficacy base recovery prognosis heal recover cure treatment gene survival therapy prognostic prophylactic clinical outcome therapeutic prognoses personalized survive healing Gleason genotype resistance, 1.0, 1.0, 1.5, 0.6, best_fields, 1.0, -1.0, 2.0, 10, phrase, 1, 1.0, 1.0, 1.0, 1, 2.0, mouse tumor pathogenesis tissue staining model case development specific cell, phrase, phrase, 1.0, 1, 1.0, 1.5])_ElasticSearchQuery

which is actually quite worse than the optimized jlpmcommon2 run. However, there are also some differences like the usage of the UMLS instead of the Lexigram API, the removal of the "match_all" clause within the disease query, more positive PM boosters (which should be selected for the optimal ones) and small differences in the fields being searched. Hopefully, the SMAC optimization will find good parameters to improve upon this.