The ElasticSearchSetup now also creates copies of the created indexes. Thus, not just `trecpm19_ct_bm25` but `trecpm19_ct_bm25_copy$i`.
This is a necessity resulting from the independent SMAC runs: In a 10-fold crossval I will have 10 independent SMAC runs sending queries to the same index. However, when one run changes the BM25 hyperparameters k1 and b, the new settings would also be valid for all other runs concurrently searching the index. This invalidates the result outcomes. Thus we need copies of the indices and must organize that no run changes the hyper parameters while another run is using the index.