Noticed that the `EvaluateConfigurationRoute` did not write the -1 relevance sample Qrels, resulting in inaccurate infNDCG estimates.
The (sample)qrel writing algorithms were confused in the sense the at places where the samples should have been used, the sample-cleaned version was used instead. This is fixed now.
Note that there are currently some dirty hacks in here to create the "infNDCG_2" measure which I used for comparison between the real infNDCG and the infNDCG computed without the samples. While there is a strong correlation, the relationship infNDCG>=infNDCG does not always hold.