Removed the reranker and the unifying field from ElasticSearch(Query). Again to free those classes from logic they don't need to perform.
Also we already have the ranker in `Experiment` which was confusing anyway. The main idea for the ElasticSearch reranker was to realise rank fusion on document hits found in different indices. But that should also be possible at another point (given that the index information is passed along).
The unifying field is now a decorator realized by `ResultByFieldValueUnificationDecorator`.